requirements.txt
boto3
botostubs
requests
hurry.filesize